refactor: reorganize file handling utilities and update references for clarity
