+++ src/core/index.ts
@@ -68,3 +68,10 @@ export type { ComputeConfig, RhinoModelUnit, RetryPolicy } from './types';
 
 // String utilities
 export { toCamelCase, camelcaseKeys } from './utils/camel-case';
+
+// ============================================================================
+// FILE HANDLING (generic zip / base64 / download — not Grasshopper-specific)
+// ============================================================================
+
+export { extractFilesFromComputeResponse, downloadFileData } from './files/handle-files';
+export type { ProcessedFile, FileData, FileBaseInfo } from './files/types';
