+++ src/features/grasshopper/index.ts
@@ -29,7 +29,7 @@ export { hashSolveInput, stableStringify, fnv1a } from './scheduler/stable-hash'
 // ============================================================================
 // COMPUTATION
 // ============================================================================
-export { solveGrasshopperDefinition } from './compute/solve';
+export { solveGrasshopperDefinition } from './solve';
 
 // ============================================================================
 // I/O PROCESSING
@@ -57,10 +57,10 @@ export { TreeBuilder } from './data-tree/data-tree';
 export type { DataTreeValue } from './data-tree/data-tree';
 
 // ============================================================================
-// FILE HANDLING
+// FILE HANDLING (generic — now lives in core/files, re-exported here for back-compat)
 // ============================================================================
-export { extractFilesFromComputeResponse, downloadFileData } from './file-handling/handle-files';
-export type { ProcessedFile, FileData, FileBaseInfo } from './file-handling/types';
+export { extractFilesFromComputeResponse, downloadFileData } from '@/core/files/handle-files';
+export type { ProcessedFile, FileData, FileBaseInfo } from '@/core/files/types';
 
 // ============================================================================
 // TYPES
